refactor(kernel): move attribution stores into write coordination
