+++ packages/kernel/src/store/index.ts
@@ -6,6 +6,6 @@ export * from "../persistence/local/local-lock-registry.ts";
 export * from "../persistence/git/local-version-control-system.ts";
 export * from "../persistence/markdown/markdown-artifact-store.ts";
 export * from "./write-journal-coordinator.ts";
-export * from "./write-journal-attribution-events.ts";
-export * from "./authority-attribution-event-v2-log.ts";
-export * from "./authority-attribution-event-v2-recovery.ts";
+export * from "../write-coordination/attribution/legacy-attribution-event-store.ts";
+export * from "../write-coordination/attribution/authority-attribution-event-v2-log.ts";
+export * from "../write-coordination/attribution/authority-attribution-event-v2-recovery.ts";
